requirements/base.txt
Django==1.11.2
psycopg2==2.7.1
pytz==2017.2
sqlparse==0.2.3
djangorestframework==3.6.3
